refactor: Rename modal component to recipemodal
