+++ recipe-app/src/CSS/styles.css
@@ -3,5 +3,5 @@
 @import url(Gallery.css);
 @import url(Card.css);
 @import url(Welcome.css);
-@import url(Modal.css);
+@import url(RecipeModal.css);
 @import url(Error.css);
